+++ CHANGELOG.md
@@ -8,6 +8,7 @@ History of all released versions and changes
 - Add a record-button to record the canvas as a video and save it on the user's computer.
 - Disable some session buttons if there is no session (disconnected).
 - NoVNC-Proxy: Update Websockify version from v0.12.0 -> v0.13.0.
+- Add GitHub-url on Dashboard for the selected example Docker-Image.
 
 ## v1.2.0
 - Fix bug with "Toggle keys" under a windows machine.